Refactored the MailApproveService by doing the following: * Putting (Mail)Message handling in a separate wrapper class * Writing a new repository method to load vacation requests without the need to mock the security * Marking it as a service and removing it from the old ApiBeansConfiguration
This reduces the lines and separates concerns a bit more.
