+++ src/test/java/de/techdev/trackr/domain/employee/vacation/support/MailApproveServiceTest.java
@@ -1,13 +1,9 @@
-package de.techdev.trackr.domain.employee.vacation;
+package de.techdev.trackr.domain.employee.vacation.support;
 
+import de.techdev.trackr.domain.employee.vacation.VacationRequest;
 import org.junit.Before;
 import org.junit.Test;
 
-import javax.mail.Session;
-import javax.mail.internet.*;
-import java.util.Properties;
-
-import static org.echocat.jomon.testing.BaseMatchers.isNull;
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
 
@@ -18,48 +14,6 @@ public class MailApproveServiceTest {
     @Before
     public void setUp() throws Exception {
         mailApproveService = new MailApproveService();
-    }
-
-    @Test
-    public void getSender() throws Exception {
-        MimeMessage mail = new MimeMessage(Session.getDefaultInstance(new Properties()));
-        mail.setFrom(new InternetAddress("test@techdev.de"));
-        String sender = mailApproveService.getSender(mail);
-        assertThat(sender, is("test@techdev.de"));
-    }
-
-    @Test
-    public void getSenderReturnsNullWhenNotTechdev() throws Exception {
-        MimeMessage mail = new MimeMessage(Session.getDefaultInstance(new Properties()));
-        mail.setFrom(new InternetAddress("test@gmail.de"));
-        String sender = mailApproveService.getSender(mail);
-        assertThat(sender, isNull());
-    }
-
-    @Test
-    public void getBodyPlaintext() throws Exception {
-        MimeMessage mail = new MimeMessage(Session.getDefaultInstance(new Properties()));
-        mail.setContent("Test", "text/plain");
-        String body = mailApproveService.extractContentAsString(mail);
-        assertThat(body, is("Test"));
-    }
-
-    @Test
-    public void getBodyMimeMultipart() throws Exception {
-        MimeMessage mail = new MimeMessage(Session.getDefaultInstance(new Properties()));
-        MimeMultipart mimeMultipart = new MimeMultipart();
-
-        MimeBodyPart plainTextPart = new MimeBodyPart(new InternetHeaders(), "Test".getBytes());
-        plainTextPart.setContent("Test", "text/plain");
-        mimeMultipart.addBodyPart(plainTextPart);
-
-        MimeBodyPart otherPart = new MimeBodyPart(new InternetHeaders(), "Other".getBytes());
-        otherPart.setContent("<b>Other</b>", "text/html");
-        mimeMultipart.addBodyPart(otherPart);
-
-        mail.setContent(mimeMultipart);
-        String body = mailApproveService.extractContentAsString(mail);
-        assertThat(body, is("Test"));
     }
 
     @Test
